Reorganise/rename to avoid import errors on various platforms. Rename probable_immunity_app folder to probable_immunity web_app. Move code from probable_immunity_app.__init__ to probable_immunity_app.app_factory so that create_app function can be explicitly imported and the return value (the app) assigned to a variable that can be passed to pythonanywhere's wsgi.
